refactor sign up and log in screens
