+++ konveksi/src/screens/LoginScreen.js
@@ -1,117 +1,30 @@
-import React, { useState, useContext } from 'react';
-import { View, Text, StyleSheet, TextInput, Dimensions, TouchableOpacity } from 'react-native';
+import React, { useContext } from 'react';
+import { View, StyleSheet } from 'react-native';
+import { NavigationEvents } from 'react-navigation';
 import { Context as AuthContext } from '../context/AuthContext';
+import AuthForm from '../components/AuthForm';
+import NavLink from '../components/NavLink';
 
 const LoginScreen = ({navigation}) => {
-  const { state, signin } = useContext(AuthContext);
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
+    const { state, signin, clearErrorMessage } = useContext(AuthContext);
 
-  const submit = () => {
-    //Fill Submit Code here
-    signin({email, password});
-    //navigation.navigate('Home');
-    return;
-  };
-
-  return (
-    <View>
-      <View style={styles.viewTitle}>
-      <Text style={styles.title}> Welcome to Konveksiin Aja </Text>
-      </View>
-      <View style={styles.viewInput}>
-      <Text style={styles.textInput}> Email Address </Text>
-      <TextInput style={styles.inputBox}
-        autoCapitalize="none"
-        autoCorrect={false}
-        value={email}
-        onChangeText={(text) => {
-          setEmail(text);
-        }}/>
-      <Text style={styles.textInput}> Password </Text>
-      <TextInput style={styles.inputBox}
-        secureTextEntry={true}
-        autoCapitalize="none"
-        autoCorrect={false}
-        value={password}
-        onChangeText={(text) => {
-          setPassword(text);
-        }}/>
-
-      {
-      state.errorMessage ?
-      <Text style={styles.errorMessage}> {state.errorMessage} </Text>
-      : null
-      }
-
-      <TouchableOpacity
-      style={styles.button}
-      onPress={() => {
-        submit();
-      }}>
-        <Text style={styles.buttonText}> Login </Text>
-      </TouchableOpacity>
-
-      <TouchableOpacity
-      onPress={() => {
-        navigation.navigate('Signup');
-      }}>
-        <Text style={styles.signin}> Don't have an account? <Text style={{color:'rgb(32,101,145)', fontWeight:'bold'}}>Sign up</Text> instead </Text>
-      </TouchableOpacity>
-      </View>
-    </View>
-  )
+      return (
+          <View>
+              <NavigationEvents onWillFocus={clearErrorMessage} />
+              <AuthForm
+                errorMessage={state.errorMessage}
+                onSubmit={signin}
+                submitButtonText="Login"
+              />
+              <NavLink
+                routeName="Signup"
+                text="Don't have an account?"
+                routeText="Sign up"
+              />
+          </View>
+      )
 };
 
-const styles = StyleSheet.create({
-  viewTitle : {
-    backgroundColor:'rgb(32,101,145)',
-    height: Dimensions.get('window').height*2.5/5,
-    justifyContent: 'center'
-  },
-  title : {
-    fontSize: 20,
-    alignSelf: 'center',
-    color: 'white',
-    fontWeight: 'bold'
-  },
-  viewInput : {
-    backgroundColor:'white',
-    height: Dimensions.get('window').height*3/5,
-    paddingTop: 30
-  },
-  textInput: {
-    alignSelf:'center',
-    fontSize: 15,
-    fontWeight: 'bold',
-    marginVertical: 20
-  },
-  inputBox: {
-    marginHorizontal:50,
-    borderBottomWidth:1
-  },
-  button: {
-    backgroundColor:'rgb(32,101,145)',
-    paddingVertical:20,
-    paddingHorizontal:50,
-    alignSelf:'center',
-    justifyContent:'center',
-    marginVertical:30,
-    borderRadius:15
-  },
-  buttonText: {
-    alignSelf:'center',
-    color:'white'
-  },
-  errorMessage: {
-    fontSize : 16,
-    color: 'red',
-    textAlign:'center',
-    marginTop: 10
-  },
-  signin: {
-    textAlign:'center'
-  },
-});
+const styles = StyleSheet.create({});
 
 export default LoginScreen;
